chore: rename files plus load env before import app
