+++ apps/examples/express/package.json
@@ -27,11 +27,12 @@
   "dependencies": {
     "@auth/core": "workspace:*",
     "@auth/express": "workspace:*",
+    "@types/pug": "^2.0.10",
     "dotenv": "^16.3.1",
-    "express": "^4.16.2",
+    "express": "^4.18.2",
     "http-errors": "^1.7.1",
     "morgan": "^1.9.1",
-    "pug": "^3.0.1",
+    "pug": "^3.0.2",
     "tailwindcss": "^3.3.3"
   },
   "devDependencies": {
